pyproject.toml
[project]
name = "cubix-data-engineer-capstone"
version = "0.1.3"
description = ""
authors = [
    {name = "Marton Szalok",email = "marton.szalok@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyspark (>=3.5.4)",
    "numpy (>=1)"
]

[project.optional-dependencies]
delta = ["delta-spark>=3.3.0"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pyarrow = "^19.0.0"
pandas = "^2.2.3"
pre-commit = "^4.5.1"

[tool.pytest.ini_options]
textpaths = ["./tests"]
filterwarnings = ["ignore:DeprecationWarning"]
addopts = "--disable-warnings -p no:warnings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

